requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary	
python-jose
passlib